requirements.txt
fastapi[standard]==0.115.0
python-socketio==5.11.14
